+++ src/Raven.Server/Documents/Indexes/Persistence/QueryFilter.cs
@@ -20,15 +20,15 @@ public class QueryFilter : IDisposable
 {
     private readonly IndexQueryServerSide _query;
     private readonly DocumentsOperationContext _documentsContext;
-    private readonly Reference<int> _skippedResults;
-    private readonly Reference<int> _scannedDocuments;
+    private readonly Reference<long> _skippedResults;
+    private readonly Reference<long> _scannedDocuments;
     private readonly IQueryResultRetriever _retriever;
     private readonly QueryTimingsScope _queryTimings;
     private readonly ScriptRunner.SingleRun _filterScriptRun;
     private ScriptRunner.ReturnRun _filterSingleRun;
 
-    public QueryFilter(Index index, IndexQueryServerSide query, DocumentsOperationContext documentsContext, Reference<int> skippedResults,
-        Reference<int> scannedDocuments, IQueryResultRetriever retriever, QueryTimingsScope queryTimings)
+    public QueryFilter(Index index, IndexQueryServerSide query, DocumentsOperationContext documentsContext, Reference<long> skippedResults,
+        Reference<long> scannedDocuments, IQueryResultRetriever retriever, QueryTimingsScope queryTimings)
     {
         _query = query;
         _documentsContext = documentsContext;
